refactor: NhClient 및 DTO들을 external.nhdevelopers 패키지로 이동 (#20)
