+++ app/src/main/java/com/mangoboss/app/external/nhdevelopers/dto/response/NhDepositorAccountNumberResponse.java
@@ -1,4 +1,4 @@
-package com.mangoboss.app.dto.payroll.nhdevelopers.response;
+package com.mangoboss.app.external.nhdevelopers.dto.response;
 
 
 public record NhDepositorAccountNumberResponse (
